requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
jinja2==3.1.4
aiofiles==24.1.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosqlite==0.20.0
python-dotenv==1.0.1
obd==0.7.1
pytest==8.3.3
pytest-asyncio==0.24.0
